backend/ml/requirements.txt
# Python ML Service - Job Portal
# FastAPI application for resume parsing and job matching

fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Database
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.25
psycopg2-binary==2.9.9

# NLP & ML
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4

# Resume parsing
pdfplumber==0.10.3
python-docx==1.1.0
pypdf==3.17.4

# Vector database
pgvector==0.2.4

# Redis
redis==5.0.1

# Utilities
httpx==0.26.0
python-dotenv==1.0.0
loguru==0.7.2
